+++ addon/controllers/edit-form.js
@@ -290,9 +290,10 @@ export default Ember.Controller.extend(Ember.Evented, FlexberryLookupMixin, Erro
 
       @method actions.close
       @param {Boolean} skipTransition If `true`, then transition during close form process will be skipped.
+      @param {Boolean} rollBackModel Flag: indicates whether to set flag to roll back model after route leave (if `true`) or not (if `false`).
     */
-    close(skipTransition) {
-      this.close(skipTransition);
+    close(skipTransition, rollBackModel) {
+      this.close(skipTransition, rollBackModel);
     },
   },
 
@@ -345,6 +346,7 @@ export default Ember.Controller.extend(Ember.Evented, FlexberryLookupMixin, Erro
     }).catch((errorData) => {
       this.set('state', 'error');
       this.onSaveActionRejected(errorData);
+      return Ember.RSVP.reject(errorData);
     }).finally((data) => {
       this.onSaveActionAlways(data);
     });
@@ -402,12 +404,13 @@ export default Ember.Controller.extend(Ember.Evented, FlexberryLookupMixin, Erro
 
     @method close
     @param {Boolean} skipTransition If `true`, then transition during close form process will be skipped.
+    @param {Boolean} rollBackModel Flag: indicates whether to set flag to roll back model after route leave (if `true`) or not (if `false`).
   */
-  close(skipTransition) {
+  close(skipTransition, rollBackModel) {
     this.set('state', '');
     this.onCloseActionStarted();
     if (!skipTransition) {
-      this.transitionToParentRoute(skipTransition);
+      this.transitionToParentRoute(skipTransition, rollBackModel);
     }
   },