+++ tests/dummy/app/controllers/components-examples/flexberry-groupedit/configurate-row-example.js
@@ -2,24 +2,6 @@ import Ember from 'ember';
 import EditFormController from 'ember-flexberry/controllers/edit-form';
 
 export default EditFormController.extend({
-  /**
-    Counter to mark created records.
-
-    @property _itemsCounter
-    @private
-    @type Number
-    @default 0
-  */
-  _itemsCounter: 0,
-
-  /**
-    Flag indicates that component have to check on model changes and display it.
-
-    @property searchForContentChange
-    @type Boolean
-    @default true
-  */
-  searchForContentChange: true,
 
   /**
     Configurate rows 'flexberry-groupedit' component by address.
@@ -29,43 +11,15 @@ export default EditFormController.extend({
   */
   configurateRowByFlag: 1,
 
-  _configurateRowByFlag: Ember.observer('configurateRowByFlag', function() {
-    let rowConfig = { customClass: '' };
-
-    this.get('records').forEach((record, index, records) => {
-      this.send('configurateRow', rowConfig, record);
-    });
-  }),
-
   /**
-    Template text for 'flexberry-groupedit' component.
+    Current records.
 
-    @property componentTemplateText
-    @type String
-  */
-  componentTemplateText: new Ember.Handlebars.SafeString(
-    '{{flexberry-groupedit<br>' +
-    '  configurateRow=(action \"configurateRow\")<br>' +
-    '}}'),
-
-  /**
-    Component settings metadata.
-
-    @property componentSettingsMetadata
+    @property _records
     @type Object[]
+    @protected
+    @readOnly
   */
-  componentSettingsMetadata: Ember.computed('i18n.locale', function() {
-    let componentSettingsMetadata = Ember.A();
-
-    componentSettingsMetadata.pushObject({
-      settingName: 'configurateRowByFlag',
-      settingType: 'boolean',
-      settingDefaultValue: 1,
-      bindedControllerPropertieName: 'configurateRowByFlag',
-    });
-
-    return componentSettingsMetadata;
-  }),
+  records: [],
 
   actions: {
     /**
@@ -77,7 +31,7 @@ export default EditFormController.extend({
       }
 
       if (record.get('flag') === this.get('configurateRowByFlag')) {
-        rowConfig.showDeleteButtonInRow = false;
+        Ember.set(rowConfig, 'canBeDeleted', false);
       }
     },
   },